% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion-function.R
\name{completion}
\alias{completion}
\alias{completion.default}
\title{Computes estimated success rate with confidence intervals}
\usage{
completion(.success, ...)

\method{completion}{default}(.success, .trials, ...)
}
\arguments{
\item{.success}{the total number of successes}

\item{...}{further arguments passed to or from other methods}

\item{.trials}{the total number of trials}
}
\value{
a dataframe with stats
}
\description{
Computes estimated success rate with confidence intervals
}
\seealso{
Other success rate estimators: \code{\link{laplace}},
  \code{\link{mle}}, \code{\link{wald.ci}},
  \code{\link{wilson}}
}
