% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.success_rate.helper-function.R
\name{df.success_rate.helper}
\alias{df.success_rate.helper}
\title{Prepare a data frame for analysis by success_rate function}
\usage{
df.success_rate.helper(.data, ...)
}
\arguments{
\item{.data}{a data frame}

\item{...}{further arguments passed to or from other methods}
}
\value{
a dataframe with stats
}
\description{
Prepare a data frame for analysis by success_rate function
}
\seealso{
Other success rate estimators: \code{\link{laplace}},
  \code{\link{mle}}, \code{\link{success_rate_fig}},
  \code{\link{success_rate}}, \code{\link{wald.ci}},
  \code{\link{wilson}}
}
\concept{success rate estimators}
