% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_mean-function.R
\name{geo_mean}
\alias{geo_mean}
\alias{geo_mean.default}
\alias{geo_mean.data.frame}
\alias{geo_mean.list}
\title{Computes the geometric mean for small sample size (n<=25) completion times}
\usage{
geo_mean(.data, ...)

\method{geo_mean}{default}(.data, ...)

\method{geo_mean}{data.frame}(.data, ...)

\method{geo_mean}{list}(.data, ...)
}
\arguments{
\item{.data}{a vector, one-column dataframe, or a one-dimensional list of completion time values}

\item{...}{further arguments passed to or from other methods}
}
\value{
a geometric mean
}
\description{
Computes the geometric mean for small sample size (n<=25) completion times
}
\concept{Completion time estimators}
