% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace-function.R
\name{laplace}
\alias{laplace}
\alias{laplace.default}
\title{Computes success rate using Laplace estimate}
\usage{
laplace(.success, ...)

\method{laplace}{default}(.success, .trials, ...)
}
\arguments{
\item{.success}{the total number of successes}

\item{...}{further arguments passed to or from other methods}

\item{.trials}{the total number of trials}
}
\value{
an estimated succes rate (%) as a numeric
}
\description{
Computes success rate using Laplace estimate
}
\seealso{
Other success rate estimators: \code{\link{df.success_rate.helper}},
  \code{\link{mle}}, \code{\link{success_rate_fig}},
  \code{\link{success_rate}}, \code{\link{wald.ci}},
  \code{\link{wilson}}
}
\concept{success rate estimators}
