% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings_summary-function.R
\name{ratings_summary}
\alias{ratings_summary}
\alias{ratings_summary.default}
\alias{ratings_summary.data.frame}
\title{Compute summary stats for ratings}
\usage{
ratings_summary(.data, ...)

\method{ratings_summary}{default}(.data, ...)

\method{ratings_summary}{data.frame}(.data, .name, ...)
}
\arguments{
\item{.data}{a vector, one-column dataframe, or a one-dimensional list of completion time values}

\item{...}{further arguments passed to or from other methods}

\item{.name}{The name of the column with ratings data (Optional)}
}
\value{
a dataframe of statistics
}
\description{
Compute summary stats for ratings
}
\concept{Data summaries time estimators}
