% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate-function.R
\name{success_rate}
\alias{success_rate}
\alias{success_rate.default}
\alias{success_rate.data.frame}
\title{Computes estimated success rate with confidence intervals for small samples, following Sauro and Lewis (2012)}
\usage{
success_rate(.success, ...)

\method{success_rate}{default}(.success, .trials, ...)

\method{success_rate}{data.frame}(.success, ...)
}
\arguments{
\item{.success}{the total number of successes}

\item{...}{further arguments passed to or from other methods}

\item{.trials}{the total number of trials}
}
\value{
a dataframe with stats
}
\description{
Computes estimated success rate with confidence intervals for small samples, following Sauro and Lewis (2012)
}
\seealso{
Other success rate estimators: \code{\link{df.success_rate.helper}},
  \code{\link{laplace}}, \code{\link{mle}},
  \code{\link{success_rate_fig}}, \code{\link{wald.ci}},
  \code{\link{wilson}}
}
\concept{success rate estimators}
