% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_fig-function.R
\name{success_rate_fig}
\alias{success_rate_fig}
\alias{success_rate_fig.default}
\title{Creates bar graph of success rates}
\usage{
success_rate_fig(.data, ...)

\method{success_rate_fig}{default}(.data, ...,
  xlabel = "Objective Success (\%)", ylabel = "Task",
  legend_lab = "Group")
}
\arguments{
\item{.data}{The results (as a data frame) from \code{success_rate} function}

\item{...}{further arguments passed to or from other methods}

\item{xlabel}{Label for the x axis}

\item{ylabel}{Label for the y axis}

\item{legend_lab}{Label for the legend (when groups are compared)}

\item{legend_lab}{Label for the legend (when groups are compared)}
}
\value{
a ggplot figure
}
\description{
Creates bar graph of success rates
}
\seealso{
Other success rate estimators: \code{\link{df.success_rate.helper}},
  \code{\link{laplace}}, \code{\link{mle}},
  \code{\link{success_rate}}, \code{\link{wald.ci}},
  \code{\link{wilson}}
}
\concept{success rate estimators}
