% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_critical-function.R
\name{t_critical}
\alias{t_critical}
\alias{t_critical.default}
\title{A convenience function based on ``qt()`` that returns a t-critical value}
\usage{
t_critical(.alpha = 0.05, ...)

\method{t_critical}{default}(.alpha = 0.05, ..., .df, .form = NULL)
}
\arguments{
\item{.alpha}{the level of significance, defaulting to 0.05}

\item{...}{further arguments passed to or from other methods}

\item{.df}{the degrees of freedom}

\item{.form}{an optional formula, passed to ``qt()``, to specify how the critical T value should be computed}
}
\value{
a t-critical value
}
\description{
A convenience function based on ``qt()`` that returns a t-critical value
}
\concept{Confidence Intervals}
