% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.ci-function.R
\name{wald.ci}
\alias{wald.ci}
\alias{wald.ci.default}
\title{Computes confidence intervals using adjusted Wald (Agresti & Coull, 1998)}
\usage{
wald.ci(.success, ...)

\method{wald.ci}{default}(.success, .trials, .Z = 1.96, ...)
}
\arguments{
\item{.success}{the proportion of successes}

\item{...}{further arguments passed to or from other methods}

\item{.trials}{the total number of trials}

\item{.Z}{z score for confidence limit}
}
\value{
a list with the lower and upper confidence limits (%) as numerics
}
\description{
Computes confidence intervals using adjusted Wald (Agresti & Coull, 1998)
}
\seealso{
Other success rate estimators: \code{\link{df.success_rate.helper}},
  \code{\link{laplace}}, \code{\link{mle}},
  \code{\link{success_rate_fig}},
  \code{\link{success_rate}}, \code{\link{wilson}}
}
\concept{success rate estimators}
