% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilson-function.R
\name{wilson}
\alias{wilson}
\alias{wilson.default}
\title{Computes success rate using Wilson estimate}
\usage{
wilson(.success, ...)

\method{wilson}{default}(.success, .trials, ...)
}
\arguments{
\item{.success}{the total number of successes}

\item{...}{further arguments passed to or from other methods}

\item{.trials}{the total number of trials}
}
\value{
an estimated succes rate (%) as a numeric
}
\description{
Computes success rate using Wilson estimate
}
\seealso{
Other success rate estimators: \code{\link{laplace}},
  \code{\link{mle}}, \code{\link{wald.ci}}
}
