% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_figure-function.R
\name{comp_figure}
\alias{comp_figure}
\alias{comp_figure.default}
\title{Creates bar graph of completion rates}
\usage{
comp_figure(.data, ...)

\method{comp_figure}{default}(.data, ...)
}
\arguments{
\item{.data}{The results (as a data frame) from \code{Completion} function}

\item{...}{further arguments passed to or from other methods}
}
\value{
a ggplot figure
}
\description{
Creates bar graph of completion rates
}
\seealso{
Other success rate estimators: \code{\link{completion}},
  \code{\link{df.completion.helper}},
  \code{\link{laplace}}, \code{\link{mle}},
  \code{\link{wald.ci}}, \code{\link{wilson}}
}
