% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.completion.helper-function.R
\name{df.completion.helper}
\alias{df.completion.helper}
\title{Prepared a data frame for analysis by completion function}
\usage{
df.completion.helper(.data, ...)
}
\arguments{
\item{.data}{a data frame}

\item{...}{further arguments passed to or from other methods}
}
\value{
a dataframe with stats
}
\description{
Prepared a data frame for analysis by completion function
}
\seealso{
Other success rate estimators: \code{\link{comp_figure}},
  \code{\link{completion}}, \code{\link{laplace}},
  \code{\link{mle}}, \code{\link{wald.ci}},
  \code{\link{wilson}}
}
