% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-function.R
\name{mle}
\alias{mle}
\alias{mle.default}
\title{Computes success rate using Maximum Liklihood estimate}
\usage{
mle(.success, ...)

\method{mle}{default}(.success, .trials, ...)
}
\arguments{
\item{.success}{the total number of successes}

\item{...}{further arguments passed to or from other methods}

\item{.trials}{the total number of trials}
}
\value{
an estimated succes rate (%) as a numeric
}
\description{
Computes success rate using Maximum Liklihood estimate
}
\seealso{
Other success rate estimators: \code{\link{df.success_rate.helper}},
  \code{\link{laplace}}, \code{\link{success_rate_fig}},
  \code{\link{success_rate}}, \code{\link{wald.ci}},
  \code{\link{wilson}}
}
\concept{success rate estimators}
